+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 There are next changes:
 
+- use mb_orig_* functions
+
 ## v2.0.0
 
 There are next changes: